Add detail map and edge generation
Added an extra world layer to keep track of - the detail map. This
layer will hold 16 x 16 details to add some extra visual detail to
the world.

To help handle this additional map layer (and any additional layers
yet to come) we've added a WorldData class. This mostly just serves
as storage for our raw and shaped world data. In the future, we may
have the WorldData class manage the sprite lists too.

As we now have 16 x 16 tiles designed to be used with 32 x 32 tiles,
the use16 functionality has been removed.

We also added our first use for the detail layer: edge passing, which
passes over the map and adds transition edges, which look quite
nice! Unfortuanetly, the process is pretty intensive in terms of
memory, so we'll need to fine tune the algorithm some more.

+++ src/city/world/data.py
@@ -2,7 +2,7 @@ import multiprocessing as mp
 import numpy
 
 # Only need these two items from ctypes, and they come with prefixes
-from ctypes import c_byte
+from ctypes import c_byte, c_int16
 
 # In order to save time on rendering tiles, we divide the map up into squares.
 # We then assign each square the most common terrain tile type there-in.
@@ -34,7 +34,7 @@ class WorldData(object):
         # flowers, or something).
         # Each terrain tile gets 4 tiles : NW, NE, SW, and SE.
         # Ergo, 16 x 16 Tiles
-        self.detail_raw = mp.Array(c_byte, (x_len * 2) * (y_len * 2))
+        self.detail_raw = mp.Array(c_int16, (x_len * 2) * (y_len * 2))
 
         # The structures that go over our terrain. We track them separately so
         # we can more easily see information SPECIFICALLY about structures.
@@ -53,7 +53,7 @@ class WorldData(object):
         a = numpy.frombuffer( self.terrain_raw.get_obj(), dtype=c_byte )
         self.terrain_shaped = a.reshape( (x_len, y_len) )
 
-        a = numpy.frombuffer( self.detail_raw.get_obj(), dtype=c_byte )
+        a = numpy.frombuffer( self.detail_raw.get_obj(), dtype=c_int16 )
         self.detail_shaped = a.reshape( (x_len * 2, y_len * 2) )
 
         a = numpy.frombuffer( self.struct_raw.get_obj(), dtype=c_byte )
